# doc/part-3/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Volunteer Information System API
  description: API для системы управления волонтерскими мероприятиями
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            read: Чтение данных
            write: Запись данных
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: Уникальный идентификатор пользователя
        organisationId:
          type: integer
          nullable: true
          description: ID организации (для представителей организации)
        role:
          type: string
          enum: [USER, ORG_REPRESENTATIVE, ADMIN]
          description: Роль пользователя
        name:
          type: string
          description: Полное имя пользователя
        email:
          type: string
          format: email
          description: Email пользователя
        description:
          type: string
          nullable: true
          description: Описание "о себе"
        birthday:
          type: string
          format: date
          nullable: true
          description: День рождения
        location:
          type: string
          nullable: true
          description: Место проживания
        volunteerHours:
          type: number
          format: float
          default: 0
          description: Количество отработанных волонтерских часов
        rating:
          type: number
          format: float
          nullable: true
          description: Средний рейтинг пользователя по отзывам
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
paths:
...
  /events/mass/{id}/participate:
    post:
      tags: [participation]
      summary: Подать заявку на участие в массовом мероприятии
      description: Авторизованный пользователь подает заявку на участие в массовом волонтерском мероприятии
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор мероприятия
      responses:
        '201':
          description: Заявка на участие успешно подана
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Мероприятие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/individual/{id}/participate:
    post:
      tags: [participation]
      summary: Подать заявку на участие в индивидуальной заявке
      description: Авторизованный пользователь подает заявку на участие в индивидуальной волонтерской заявке
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор заявки
      responses:
        '201':
          description: Заявка на участие успешно подана
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заявка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /events/mass/{id}/participants:
    get:
      tags: [participation]
      summary: Получение списка участников массового мероприятия
      description: Автор мероприятия или представители организации получают список участников
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор мероприятия
      responses:
        '200':
          description: Список участников успешно получен
          content:
            application/json:
              schema:
                type: object
                properties:
                  participants:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  totalCount:
                    type: integer
                    description: Общее количество участников
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен (пользователь не является автором/представителем организации)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Мероприятие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/individual/{id}/participants:
    get:
      tags: [participation]
      summary: Получение списка участников индивидуальной заявки
      description: Автор заявки получает список участников
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор заявки
      responses:
        '200':
          description: Список участников успешно получен
          content:
            application/json:
              schema:
                type: object
                properties:
                  participants:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  totalCount:
                    type: integer
                    description: Общее количество участников
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен (пользователь не является автором)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заявка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/mass/{id}/requests:
    get:
      tags: [participation]
      summary: Получение списка запросов на участие в массовом мероприятии
      description: Представители организации получают список запросов на участие для принятия/отклонения
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор мероприятия
      responses:
        '200':
          description: Список запросов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    requestId:
                      type: integer
                    user:
                      $ref: '#/components/schemas/User'
                    applicationDate:
                      type: string
                      format: date-time
                    status:
                      type: string
                      enum: [PENDING, ACCEPTED, REJECTED]
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Мероприятие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/individual/{id}/requests:
    get:
      tags: [participation]
      summary: Получение списка запросов на участие в индивидуальной заявке
      description: Автор заявки получает список запросов на участие для принятия/отклонения
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор заявки
      responses:
        '200':
          description: Список запросов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    requestId:
                      type: integer
                    user:
                      $ref: '#/components/schemas/User'
                    applicationDate:
                      type: string
                      format: date-time
                    status:
                      type: string
                      enum: [PENDING, ACCEPTED, REJECTED]
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заявка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/mass/{eventId}/requests/{requestId}/accept:
    post:
      tags: [participation]
      summary: Принятие запроса на участие в массовом мероприятии
      description: Представитель организации принимает запрос волонтера на участие
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор мероприятия
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор запроса
      responses:
        '200':
          description: Запрос успешно принят
        '400':
          description: Ошибка (например, достигнут лимит участников)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Мероприятие или запрос не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/mass/{eventId}/requests/{requestId}/reject:
    post:
      tags: [participation]
      summary: Отклонение запроса на участие в массовом мероприятии
      description: Представитель организации отклоняет запрос волонтера на участие
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор мероприятия
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор запроса
      responses:
        '200':
          description: Запрос успешно отклонен
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Мероприятие или запрос не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/individual/{eventId}/requests/{requestId}/accept:
    post:
      tags: [participation]
      summary: Принятие запроса на участие в индивидуальной заявке
      description: Автор заявки принимает запрос волонтера на участие
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор заявки
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор запроса
      responses:
        '200':
          description: Запрос успешно принят
        '400':
          description: Ошибка (например, достигнут лимит участников)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заявка или запрос не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /events/individual/{eventId}/requests/{requestId}/reject:
    post:
      tags: [participation]
      summary: Отклонение запроса на участие в индивидуальной заявке
      description: Автор заявки отклоняет запрос волонтера на участие
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор заявки
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор запроса
      responses:
        '200':
          description: Запрос успешно отклонен
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заявка или запрос не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Reviews
  /events/individual/{eventId}/participants/{userId}/review:
    post:
      tags: [participation]
      summary: Оставление отзыва об участнике индивидуальной заявки
      description: Автор индивидуальной заявки оставляет отзыв о волонтере после выполнения работы
      security:
        - bearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор заявки
        - name: userId
          in: path
          required: true
          schema:
            type: integer
          description: Уникальный идентификатор волонтера
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rating
                - content
              properties:
                rating:
                  type: number
                  format: float
                  minimum: 1
                  maximum: 5
                  description: Оценка от 1 до 5
                content:
                  type: string
                  description: Текст отзыва
      responses:
        '201':
          description: Отзыв успешно добавлен
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Доступ запрещен (пользователь не автор заявки или заявка не завершена)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заявка или пользователь не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Leaderboard
...